Refactor Recent Project cards, add random tiles
Replaced the exising project cards for the Bootstrap component, this
fixed the bug with the gutters not working correctly.
Added a transform: size hover effect, and replaced placeholders for real
examples, project descriptions still pending.

Added a function to randomize the rotation for the background tiles, and
a new button to randomize it again.